+++ DropInDemo/Demo Base/BraintreeDemoAppDelegate.m
@@ -104,6 +104,16 @@ NSString *BraintreeDemoAppDelegatePaymentsURLScheme = @"com.braintreepayments.Dr
         [[NSUserDefaults standardUserDefaults] setInteger:BTFormFieldRequired forKey:@"BraintreeDemoCardholderNameSetting"];
     }
 
+    [[NSUserDefaults standardUserDefaults] setBool:NO forKey:@"BraintreeDemoAllowVaultCardOverrideSetting"];
+    if ([[[NSProcessInfo processInfo] arguments] containsObject:@"-SaveCardToggleVisible"]) {
+        [[NSUserDefaults standardUserDefaults] setBool:YES forKey:@"BraintreeDemoAllowVaultCardOverrideSetting"];
+    }
+
+    [[NSUserDefaults standardUserDefaults] setBool:YES forKey:@"BraintreeDemoVaultCardSetting"];
+    if ([[[NSProcessInfo processInfo] arguments] containsObject:@"-VaultCardIsFalse"]) {
+        [[NSUserDefaults standardUserDefaults] setBool:NO forKey:@"BraintreeDemoVaultCardSetting"];
+    }
+
     [[NSUserDefaults standardUserDefaults] removeObjectForKey:@"BraintreeTest_ForceVenmoDisplay"];
     if ([[[NSProcessInfo processInfo] arguments] containsObject:@"-ForceVenmo"]) {
         [BTDropInOverrides setDisplayVenmoOption:YES];