+++ UnitTests/BTDropInRequestTests.m
@@ -22,6 +22,8 @@
     originalRequest.cardholderNameSetting = BTFormFieldOptional;
     originalRequest.shouldMaskSecurityCode = YES;
     originalRequest.vaultManager = YES;
+    originalRequest.vaultCard = NO;
+    originalRequest.allowVaultCardOverride = YES;
 
     BTDropInRequest *copiedRequest = [originalRequest copy];
 
@@ -35,6 +37,8 @@
     XCTAssertEqual(originalRequest.cardholderNameSetting, copiedRequest.cardholderNameSetting);
     XCTAssertEqual(originalRequest.shouldMaskSecurityCode, copiedRequest.shouldMaskSecurityCode);
     XCTAssertEqual(originalRequest.vaultManager, copiedRequest.vaultManager);
+    XCTAssertEqual(originalRequest.vaultCard, copiedRequest.vaultCard);
+    XCTAssertEqual(originalRequest.allowVaultCardOverride, copiedRequest.allowVaultCardOverride);
 }
 
 @end